Moved legend interaction into Legend class.
